refactor: controller, service 구조 리팩터링
