+++ backend/src/main/java/com/HUFS19/backend/service/TagService.java
@@ -1,9 +1,13 @@
 package com.HUFS19.backend.service;
 
 import com.HUFS19.backend.repository.tag.TagRepository;
+import jakarta.transaction.Transactional;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Service
+@Transactional
 public class TagService {
     private final TagRepository tagRepository;
